Refactor timeline architecture.
