+++ packages/devtools/lib/src/timeline/timeline_screen.dart
@@ -5,7 +5,6 @@
 import 'dart:convert';
 
 import 'package:split/split.dart' as split;
-import 'package:vm_service_lib/vm_service_lib.dart' hide TimelineEvent;
 
 import '../framework/framework.dart';
 import '../globals.dart';
@@ -18,11 +17,11 @@ import '../ui/material_icons.dart';
 import '../ui/primer.dart';
 import '../ui/theme.dart';
 import '../ui/ui_utils.dart';
-import '../vm_service_wrapper.dart';
 import 'event_details.dart';
-import 'frame_flame_chart.dart';
+import 'frame_events_chart.dart';
 import 'frames_bar_chart.dart';
 import 'timeline_controller.dart';
+import 'timeline_model.dart';
 import 'timeline_protocol.dart';
 
 // Light mode is Light Blue 50 palette and Dark mode is Blue 50 palette.
@@ -78,8 +77,7 @@ const Icon _exportTimelineIcon = MaterialIcon(
 
 const Icon _clear = MaterialIcon('block', defaultButtonIconColor);
 
-const Icon _exitSnapshotModeIcon =
-    MaterialIcon('clear', defaultButtonIconColor);
+const Icon _exitIcon = MaterialIcon('clear', defaultButtonIconColor);
 
 class TimelineScreen extends Screen {
   TimelineScreen({bool disabled, String disabledTooltip})
@@ -89,18 +87,13 @@ class TimelineScreen extends Screen {
           iconClass: 'octicon-pulse',
           disabled: disabled,
           disabledTooltip: disabledTooltip,
-        ) {
-    onLoadTimelineSnapshot.listen((snapshot) {
-      _updateButtonStates();
-      _clearTimeline();
-    });
-  }
+        );
 
   TimelineController timelineController = TimelineController();
 
   FramesBarChart framesBarChart;
 
-  FrameFlameChart flameChart;
+  FrameEventsChart frameEventsChart;
 
   EventDetails eventDetails;
 
@@ -112,7 +105,7 @@ class TimelineScreen extends Screen {
 
   PButton exportButton;
 
-  PButton exitSnapshotModeButton;
+  PButton exitOfflineModeButton;
 
   CoreElement upperButtonSection;
 
@@ -152,16 +145,16 @@ class TimelineScreen extends Screen {
       ..small()
       ..clazz('margin-left')
       ..setAttribute('title', 'Clear timeline')
-      ..click(_clearTimeline);
+      ..click(clearTimeline);
 
-    exitSnapshotModeButton = PButton.icon(
-      'Exit snapshot mode',
-      _exitSnapshotModeIcon,
+    exitOfflineModeButton = PButton.icon(
+      'Exit offline mode',
+      _exitIcon,
     )
       ..small()
-      ..setAttribute('title', 'Exit snapshot mode to connect to a VM Service.')
+      ..setAttribute('title', 'Exit offline mode to connect to a VM Service.')
       ..setAttribute('hidden', 'true')
-      ..click(_exitSnapshotMode);
+      ..click(_exitOfflineMode);
 
     upperButtonSection = div(c: 'section')
       ..layoutHorizontal()
@@ -172,7 +165,7 @@ class TimelineScreen extends Screen {
             resumeButton,
           ]),
         clearButton,
-        exitSnapshotModeButton,
+        exitOfflineModeButton,
         div()..flex(),
         debugButtonSection = div(c: 'btn-group'),
         exportButton,
@@ -188,56 +181,23 @@ class TimelineScreen extends Screen {
         ..layoutVertical()
         ..flex()
         ..add(<CoreElement>[
-          flameChart = FrameFlameChart()..attribute('hidden'),
-          eventDetails = EventDetails()..attribute('hidden'),
+          frameEventsChart = FrameEventsChart(timelineController)
+            ..attribute('hidden'),
+          eventDetails = EventDetails(timelineController)..attribute('hidden'),
         ]),
     ]);
 
-    serviceManager.onConnectionAvailable.listen(_handleConnectionStart);
-    if (serviceManager.hasConnection) {
-      _handleConnectionStart(serviceManager.service);
-    }
-    serviceManager.onConnectionClosed.listen(_handleConnectionStop);
-
-    framesBarChart.onSelectedFrame.listen((TimelineFrame frame) {
-      if (frame != null && timelineController.hasStarted) {
-        flameChart.attribute('hidden', frame == null);
-        eventDetails.attribute('hidden', frame == null);
-
-        if (debugTimeline && frame != null) {
-          final buf = StringBuffer();
-          buf.writeln('UI timeline event for frame ${frame.id}:');
-          frame.uiEventFlow.format(buf, '  ');
-          buf.writeln('\nUI trace for frame ${frame.id}');
-          frame.uiEventFlow.writeTraceToBuffer(buf);
-          buf.writeln('\nGPU timeline event frame ${frame.id}:');
-          frame.gpuEventFlow.format(buf, '  ');
-          buf.writeln('\nGPU trace for frame ${frame.id}');
-          frame.gpuEventFlow.writeTraceToBuffer(buf);
-          print(buf.toString());
-        }
-
-        flameChart.update(frame);
-        eventDetails.reset();
-
-        _configureSplitter();
-      }
-    });
-
-    onSelectedFrameFlameChartItem.listen((FrameFlameChartItem item) async {
-      final TimelineEvent event = item.event;
-      ga.select(
-        ga.timeline,
-        event.isGpuEvent ? ga.timelineFlameGpu : ga.timelineFlameUi,
-        event.time.duration.inMicroseconds, // No inMilliseconds loses precision
-      );
-
-      await eventDetails.update(item);
-    });
+    _initListeners();
 
     maybeShowDebugWarning(framework);
 
     return screenDiv;
+  }
+
+  void _initListeners() {
+    timelineController.onSelectedFrame.listen((_) => _configureSplitter());
+    timelineController.onLoadOfflineData
+        .listen((_) => frameEventsChart.attribute('hidden', true));
   }
 
   void _configureSplitter() {
@@ -245,11 +205,11 @@ class TimelineScreen extends Screen {
     // already.
     if (!splitterConfigured) {
       splitter = split.flexSplit(
-        [flameChart.element, eventDetails.element],
+        [frameEventsChart.element, eventDetails.element],
         horizontal: false,
         gutterSize: defaultSplitterWidth,
         sizes: [75, 25],
-        minSize: [60, 160],
+        minSize: [50, 160],
       );
       splitterConfigured = true;
     }
@@ -275,39 +235,19 @@ class TimelineScreen extends Screen {
     _updateButtonStates();
   }
 
-  void _exitSnapshotMode() {
-    // This needs to be called first because [framework.exitSnapshotMode()] will
+  void _exitOfflineMode() {
+    // This needs to be called first because [framework.exitOfflineMode()] will
     // remove all elements from the dom if we are not connected to an app.
     // Performing operations from [_clearTimeline()] on elements that have been
     // removed will throw exceptions, so we need to maintain this order.
-    _clearTimeline();
-    eventDetails.clearCurrentSnapshot();
-    timelineController.exitSnapshotMode();
+    clearTimeline();
+    eventDetails.reset(hide: true);
+    timelineController.exitOfflineMode();
     // This needs to be called before we update the button states because it
-    // changes the value of [snapshotMode], which the button states depend on.
-    framework.exitSnapshotMode();
+    // changes the value of [offlineMode], which the button states depend on.
+    framework.exitOfflineMode();
     _updateButtonStates();
     _destroySplitter();
-  }
-
-  void _handleConnectionStart(VmServiceWrapper service) {
-    serviceManager.service.setFlag('profile_period', '50');
-    serviceManager.service.onEvent('Timeline').listen((Event event) {
-      final List<dynamic> list = event.json['timelineEvents'];
-      final List<Map<String, dynamic>> events =
-          list.cast<Map<String, dynamic>>();
-
-      if (!snapshotMode && !_manuallyPaused && !timelineController.paused) {
-        for (Map<String, dynamic> json in events) {
-          final TraceEvent e = TraceEvent(json);
-          timelineController.timelineData?.processTraceEvent(e);
-        }
-      }
-    });
-  }
-
-  void _handleConnectionStop(dynamic event) {
-    timelineController = null;
   }
 
   void _pauseRecording() {
@@ -329,60 +269,43 @@ class TimelineScreen extends Screen {
   void _updateButtonStates() {
     pauseButton
       ..disabled = _manuallyPaused
-      ..attribute('hidden', snapshotMode);
+      ..attribute('hidden', offlineMode);
     resumeButton
       ..disabled = !_manuallyPaused
-      ..attribute('hidden', snapshotMode);
-    clearButton.attribute('hidden', snapshotMode);
-    exportButton.attribute('hidden', snapshotMode);
-    exitSnapshotModeButton.attribute('hidden', !snapshotMode);
+      ..attribute('hidden', offlineMode);
+    clearButton.attribute('hidden', offlineMode);
+    exportButton.attribute('hidden', offlineMode);
+    exitOfflineModeButton.attribute('hidden', !offlineMode);
   }
 
   void _updateListeningState() async {
-    await serviceManager.serviceAvailable.future;
-
     final bool shouldBeRunning = !_manuallyPaused && isCurrentScreen;
     final bool isRunning = !timelineController.paused;
-
-    if (shouldBeRunning && isRunning && !timelineController.hasStarted) {
-      await timelineController.startTimeline();
-    }
-
-    if (shouldBeRunning && !isRunning) {
-      timelineController.resume();
-
-      await serviceManager.service
-          .setVMTimelineFlags(<String>['GC', 'Dart', 'Embedder']);
-    } else if (!shouldBeRunning && isRunning) {
-      // TODO(devoncarew): turn off the events
-      await serviceManager.service.setVMTimelineFlags(<String>[]);
-      timelineController.pause();
-    }
+    timelineController.timelineService.updateListeningState(
+      shouldBeRunning: shouldBeRunning,
+      isRunning: isRunning,
+    );
   }
 
-  void _clearTimeline() {
-    timelineTraceEvents.clear();
+  void clearTimeline() {
     debugHandledTraceEvents.clear();
     debugFrameTracking.clear();
+    timelineController.timelineData.clear();
     framesBarChart.frameUIgraph.reset();
-    flameChart.attribute('hidden', true);
-    eventDetails.attribute('hidden', true);
-    eventDetails.reset();
+    frameEventsChart.attribute('hidden', true);
+    eventDetails.reset(hide: true);
     _destroySplitter();
   }
 
   void _exportTimeline() {
     // TODO(kenzie): add analytics for this. It would be helpful to know how
     // complex the problems are that users are trying to solve.
-    final snapshot = TimelineSnapshot.from(
-      timelineTraceEvents,
-      eventDetails.cpuProfileData?.cpuProfileResponse?.json,
-      eventDetails.event,
-    );
+    final String encodedTimelineData =
+        jsonEncode(timelineController.timelineData.json);
     final now = DateTime.now();
     final timestamp =
         '${now.year}_${now.month}_${now.day}-${now.microsecondsSinceEpoch}';
-    downloadFile(snapshot.encodedJson, 'timeline_$timestamp.json');
+    downloadFile(encodedTimelineData, 'timeline_$timestamp.json');
   }
 
   /// Adds a button to the timeline that will dump debug information to text
@@ -408,52 +331,46 @@ class TimelineScreen extends Screen {
             'frame_tracking_output.txt',
           );
 
+          final timelineProtocol = timelineController.timelineProtocol;
+
           // Current status of our frame tracking elements (i.e. pendingEvents,
           // pendingFrames).
           final buf = StringBuffer();
           buf.writeln('Pending events: '
-              '${timelineController.timelineData.pendingEvents.length}');
-          for (TimelineEvent event
-              in timelineController.timelineData.pendingEvents) {
+              '${timelineProtocol.pendingEvents.length}');
+          for (TimelineEvent event in timelineProtocol.pendingEvents) {
             event.format(buf, '    ');
             buf.writeln();
           }
           buf.writeln('\nPending frames: '
-              '${timelineController.timelineData.pendingFrames.length}');
-          for (TimelineFrame frame
-              in timelineController.timelineData.pendingFrames.values) {
+              '${timelineProtocol.pendingFrames.length}');
+          for (TimelineFrame frame in timelineProtocol.pendingFrames.values) {
             buf.writeln('${frame.toString()}');
           }
-          if (timelineController
-                  .timelineData.currentEventNodes[TimelineEventType.ui.index] !=
+          if (timelineProtocol.currentEventNodes[TimelineEventType.ui.index] !=
               null) {
             buf.writeln('\nCurrent UI event node:');
-            timelineController
-                .timelineData.currentEventNodes[TimelineEventType.ui.index]
+            timelineProtocol.currentEventNodes[TimelineEventType.ui.index]
                 .format(buf, '   ');
           }
-          if (timelineController.timelineData
-                  .currentEventNodes[TimelineEventType.gpu.index] !=
+          if (timelineProtocol.currentEventNodes[TimelineEventType.gpu.index] !=
               null) {
             buf.writeln('\n Current GPU event node:');
-            timelineController
-                .timelineData.currentEventNodes[TimelineEventType.gpu.index]
+            timelineProtocol.currentEventNodes[TimelineEventType.gpu.index]
                 .format(buf, '   ');
           }
-          if (timelineController
-              .timelineData.heaps[TimelineEventType.ui.index].isNotEmpty) {
+          if (timelineProtocol.heaps[TimelineEventType.ui.index].isNotEmpty) {
             buf.writeln('\nUI heap');
-            for (TraceEventWrapper wrapper in timelineController
-                .timelineData.heaps[TimelineEventType.ui.index]
+            for (TraceEventWrapper wrapper in timelineProtocol
+                .heaps[TimelineEventType.ui.index]
                 .toList()) {
               buf.writeln(wrapper.event.json.toString());
             }
           }
-          if (timelineController
-              .timelineData.heaps[TimelineEventType.gpu.index].isNotEmpty) {
+          if (timelineProtocol.heaps[TimelineEventType.gpu.index].isNotEmpty) {
             buf.writeln('\nGPU heap');
-            for (TraceEventWrapper wrapper in timelineController
-                .timelineData.heaps[TimelineEventType.gpu.index]
+            for (TraceEventWrapper wrapper in timelineProtocol
+                .heaps[TimelineEventType.gpu.index]
                 .toList()) {
               buf.writeln(wrapper.event.json.toString());
             }
