+++ packages/devtools/lib/src/ui/ui_utils.dart
@@ -144,7 +144,7 @@ CoreElement createHotRestartButton(Framework framework) {
 }
 
 Future<void> maybeShowDebugWarning(Framework framework) async {
-  if (!snapshotMode &&
+  if (!offlineMode &&
       serviceManager.connectedApp != null &&
       !await serviceManager.connectedApp.isProfileBuild) {
     framework.showWarning(children: <CoreElement>[
