+++ packages/devtools/pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
   intl: ^0.15.0
   js: ^0.6.1+1
   meta: ^1.1.0
+  mockito: 4.0.0
   path: ^1.6.0
   pedantic: ^1.4.0
   platform_detect: ^1.3.5
