+++ packages/devtools/test/timeline_controller_test.dart
@@ -1,160 +1,114 @@
 // Copyright 2019 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-
 @TestOn('vm')
-import 'dart:convert';
-
 import 'package:devtools/src/timeline/timeline_controller.dart';
-import 'package:devtools/src/timeline/timeline_protocol.dart';
+import 'package:devtools/src/timeline/timeline_model.dart';
 import 'package:test/test.dart';
 
-import 'support/flutter_test_driver.dart' show FlutterRunConfiguration;
+import 'support/flutter_test_driver.dart';
 import 'support/flutter_test_environment.dart';
-import 'support/test_utils.dart';
+import 'support/timeline_test_data.dart';
 
 void main() {
   group('TimelineController', () {
-    final timelineController = TimelineController();
+    TimelineController timelineController;
 
     final env = FlutterTestEnvironment(
       const FlutterRunConfiguration(withDebugger: true),
     );
 
     env.afterNewSetup = () async {
-      await timelineController.startTimeline();
+      timelineController = TimelineController();
+      await timelineController.timelineService.startTimeline();
     };
 
-    // TODO(kenzie): uncomment these methods once we have proper data from the
-    //  engine or have a way to handle events from the test device.
+    tearDownAll(() async {
+      await env.tearDownEnvironment(force: true);
+    });
 
-//    tearDownAll(() async {
-//      await env.tearDownEnvironment(force: true);
-//    });
-//
-//    test('timeline data not empty', () async {
-//      await env.setupEnvironment();
-//      expect(timelineController.timelineData.frames, isNotEmpty);
-//      expect(timelineController.timelineData.frames, isNotEmpty);
-//      await env.tearDownEnvironment();
-//    });
+    test('selection', () async {
+      await env.setupEnvironment();
 
-    // TODO(kenzie): add more tests. We will be able to once we have the proper
-    //  data from the engine, allowing us to distinguish cpu events from gpu
-    //  events as well as to know accurate frame start times, end times, and
-    //  events.
+      // Select a frame.
+      final frame_0 = TimelineFrame('id_0');
+      expect(timelineController.timelineData.selectedFrame, isNull);
+      timelineController.selectFrame(frame_0);
+      expect(timelineController.timelineData.selectedFrame, equals(frame_0));
+
+      // Select a timeline event.
+      expect(timelineController.timelineData.selectedEvent, isNull);
+      expect(timelineController.timelineData.cpuProfileData, isNull);
+      timelineController.selectTimelineEvent(vsyncEvent);
+      expect(timelineController.timelineData.selectedEvent, equals(vsyncEvent));
+      expect(timelineController.timelineData.cpuProfileData, isNotNull);
+
+      // Select a different frame.
+      final frame_1 = TimelineFrame('id_1');
+      timelineController.selectFrame(frame_1);
+      expect(timelineController.timelineData.selectedFrame, equals(frame_1));
+      expect(timelineController.timelineData.selectedEvent, isNull);
+      expect(timelineController.timelineData.cpuProfileData, isNull);
+
+      await env.tearDownEnvironment();
+    });
+
+    test('add frame', () async {
+      await env.setupEnvironment();
+      expect(timelineController.timelineData.frames, isEmpty);
+      timelineController.addFrame(TimelineFrame('id'));
+      expect(timelineController.timelineData.frames.length, equals(1));
+      await env.tearDownEnvironment();
+    });
+
+    test('recordTraceForTimelineEvent', () async {
+      await env.setupEnvironment();
+
+      expect(timelineController.timelineData.traceEvents, isEmpty);
+      timelineController.recordTraceForTimelineEvent(goldenUiTimelineEvent);
+      expect(
+        timelineController.timelineData.traceEvents,
+        equals([
+          vsyncJson,
+          animatorBeginFrameJson,
+          frameworkWorkloadJson,
+          engineBeginFrameJson,
+          frameJson,
+          animateJson,
+          layoutJson,
+          buildJson,
+          compositingBitsJson,
+          paintJson,
+          compositingJson,
+          semanticsJson,
+          finalizeTreeJson,
+          endEngineBeginFrameJson,
+          endFrameworkWorkloadJson,
+          endAnimatorBeginFrameJson,
+          endVsyncJson,
+        ]),
+      );
+
+      await env.tearDownEnvironment();
+    });
+
+    test('loadOfflineData', () async {
+      await env.setupEnvironment();
+
+      final offlineData = OfflineTimelineData.parse(offlineTimelineDataJson);
+      timelineController.loadOfflineData(offlineData);
+      expect(timelineController.timelineData, equals(offlineData));
+      expect(timelineController.offlineTimelineData, equals(offlineData));
+      expect(
+        timelineController.timelineProtocol.uiThreadId,
+        equals(testUiThreadId),
+      );
+      expect(
+        timelineController.timelineProtocol.gpuThreadId,
+        equals(testGpuThreadId),
+      );
+
+      await env.tearDownEnvironment();
+    });
   }, tags: 'useFlutterSdk');
-
-  group('TimelineSnapshot', () {
-    test('init from objects', () {
-      TimelineSnapshot snapshot = TimelineSnapshot.from(null, null, null);
-      expect(snapshot.traceEvents, equals([]));
-      expect(snapshot.cpuProfile, equals({}));
-      expect(snapshot.selectedEvent, null);
-      expect(
-        snapshot.encodedJson,
-        equals('{"traceEvents":[],"cpuProfile":{},"selectedEvent":{},'
-            '"dartDevToolsScreen":"timeline"}'),
-      );
-
-      final timelineEvent = testTimelineEvent({
-        'name': 'VSYNC',
-        'cat': 'Embedder',
-        'tid': 1,
-        'pid': 94955,
-        'ts': 118039650802,
-        'ph': 'B',
-        'args': {}
-      })
-        ..time.end = const Duration(microseconds: 118039652422)
-        ..type = TimelineEventType.ui;
-      snapshot = TimelineSnapshot.from(
-        [
-          {'name': 'FakeTraceEvent'}
-        ],
-        {'type': '_CpuProfileTimelineFakeResponse'},
-        timelineEvent,
-      );
-      expect(
-        snapshot.traceEvents,
-        equals([
-          {'name': 'FakeTraceEvent'}
-        ]),
-      );
-      expect(snapshot.cpuProfile,
-          equals({'type': '_CpuProfileTimelineFakeResponse'}));
-      expect(snapshot.selectedEvent, isA<TimelineEventSnapshot>());
-      expect(
-        snapshot.selectedEvent.json,
-        equals({
-          'name': timelineEvent.name,
-          'startMicros': timelineEvent.time.start.inMicroseconds,
-          'durationMicros': timelineEvent.time.duration.inMicroseconds
-        }),
-      );
-      expect(
-        snapshot.encodedJson,
-        equals('{"traceEvents":[{"name":"FakeTraceEvent"}],'
-            '"cpuProfile":{"type":"_CpuProfileTimelineFakeResponse"},'
-            '"selectedEvent":{"name":"VSYNC","startMicros":118039650802,'
-            '"durationMicros":1620},"dartDevToolsScreen":"timeline"}'),
-      );
-    });
-
-    test('parse from json', () {
-      TimelineSnapshot snapshot = TimelineSnapshot.parse(jsonDecode(jsonEncode({
-        'traceEvents': [],
-        'cpuProfile': {},
-        'selectedEvent': {},
-        'dartDevToolsScreen': 'timeline'
-      })));
-      expect(snapshot.traceEvents, equals([]));
-      expect(snapshot.cpuProfile, equals({}));
-      expect(snapshot.selectedEvent, isNull);
-
-      snapshot = TimelineSnapshot.parse(jsonDecode(jsonEncode({
-        'traceEvents': [
-          {'name': 'FakeTraceEvent'}
-        ],
-        'cpuProfile': {'type': '_CpuProfileTimelineFakeResponse'},
-        'selectedEvent': {
-          'name': 'VSYNC',
-          'startMicros': 118039650802,
-          'durationMicros': 1620
-        },
-        'dartDevToolsScreen': 'timeline'
-      })));
-      expect(
-        snapshot.traceEvents,
-        equals([
-          {'name': 'FakeTraceEvent'}
-        ]),
-      );
-      expect(
-        snapshot.cpuProfile,
-        equals({'type': '_CpuProfileTimelineFakeResponse'}),
-      );
-      expect(snapshot.selectedEvent, isA<TimelineEventSnapshot>());
-      expect(
-        snapshot.selectedEvent.json,
-        equals({
-          'name': 'VSYNC',
-          'startMicros': 118039650802,
-          'durationMicros': 1620
-        }),
-      );
-    });
-
-    test('TimelineEventSnapshot', () {
-      final eventSnapshot = TimelineEventSnapshot('Fake event', 10, 20);
-      expect(
-        eventSnapshot.json,
-        equals({
-          'name': 'Fake event',
-          'startMicros': 10,
-          'durationMicros': 20,
-        }),
-      );
-    });
-  });
 }
